Allow waiting for all server requests to finish
This adds a thread-safe counter to Capybara::Server which keeps track of
the number of requests currently being processed by
Capybara::Server::Middleware. It exposes public methods
Server#pending_requests?, which is truthy when there are requests still
being processed, and Server#wait_for_pending_requests, which will wait
up to 60 seconds for requests to finish.